creator/requirements.txt
fastapi==0.115.6
faststream==0.5.33
aio_pika==9.5.4
pydantic==2.10.4
black==24.10.0
ruff==0.8.4
mypy==1.14.0
uvicorn==0.34.0
python-dotenv==1.0.1
